Fix pg pool usage
