+++ api/src/controllers/namespace.ts
@@ -20,14 +20,14 @@
 
 import { errorWithCode, logger } from '@bcgov/common-nodejs-utils';
 import { Response } from 'express';
-import config from '../config';
 import DataManager from '../db';
+import shared from '../libs/shared';
 import { validateObjProps } from '../libs/utils';
 
 export const createNamespace = async (
   { params, body }: { params: any, body: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { NamespaceModel } = dm;
   const { profileId } = params;
   const aBody = { ...body, profileId };
@@ -52,7 +52,7 @@ export const createNamespace = async (
 export const fetchProfileNamespaces = async (
   { params }: { params: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { NamespaceModel } = dm;
   const { profileId } = params;
 
@@ -71,7 +71,7 @@ export const fetchProfileNamespaces = async (
 export const fetchProfileNamespace = async (
   { params }: { params: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { NamespaceModel } = dm;
   const { namespaceId } = params;
 
@@ -90,7 +90,7 @@ export const fetchProfileNamespace = async (
 export const updateProfileNamespace = async (
   { params, body }: { params: any, body: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { NamespaceModel } = dm;
   const { profileId, namespaceId } = params;
   const { name, clusterId } = body;
@@ -116,7 +116,7 @@ export const updateProfileNamespace = async (
 export const archiveProfileNamespace = async (
   { params }: { params: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { NamespaceModel } = dm;
   const { profileId, namespaceId } = params;
