+++ api/src/controllers/profile.ts
@@ -20,12 +20,12 @@
 
 import { errorWithCode, logger } from '@bcgov/common-nodejs-utils';
 import { Request, Response } from 'express';
-import config from '../config';
 import DataManager from '../db';
+import shared from '../libs/shared';
 import { validateObjProps } from '../libs/utils';
 
 export const fetchAllProjectProfiles = async (req: Request, res: Response): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { ProfileModel } = dm;
 
   try {
@@ -43,7 +43,7 @@ export const fetchAllProjectProfiles = async (req: Request, res: Response): Prom
 export const fetchProjectProfile = async (
   { params }: { params: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { ProfileModel } = dm;
   const { profileId } = params;
 
@@ -62,7 +62,7 @@ export const fetchProjectProfile = async (
 export const createProjectProfile = async (
   { body }: { body: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { ProfileModel } = dm;
 
   const rv = validateObjProps(ProfileModel.requiredFields, body);
@@ -85,7 +85,7 @@ export const createProjectProfile = async (
 export const updateProjectProfile = async (
   { params, body }: { params: any, body: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { ProfileModel } = dm;
   const { profileId } = params;
   const {
@@ -125,7 +125,7 @@ export const updateProjectProfile = async (
 export const archiveProjectProfile = async (
   { params }: { params: any }, res: Response
 ): Promise<void> => {
-  const dm = new DataManager(config);
+  const dm = new DataManager(shared.pgPool);
   const { ProfileModel } = dm;
   const { profileId } = params;
